ex2

+++ src/fluorescence/world/ingameobject.cpp
@@ -527,20 +527,11 @@ const ui::WorldRenderData& IngameObject::getWorldRenderData() const {
 void IngameObject::updateGumpTextureProvider() {
 }
 
-void IngameObject::setRenderDepth(unsigned int depth) {
+void IngameObject::setRenderDepth(float depth) {
     renderDepth_ = depth;
-
-    //float dept = ((float)depth) / 20000.f;
-
-    //worldRenderData_.vertexCoordinates_[0].z = dept;
-    //worldRenderData_.vertexCoordinates_[1].z = dept;
-    //worldRenderData_.vertexCoordinates_[2].z = dept;
-    //worldRenderData_.vertexCoordinates_[3].z = dept;
-    //worldRenderData_.vertexCoordinates_[4].z = dept;
-    //worldRenderData_.vertexCoordinates_[5].z = dept;
 }
 
-unsigned int IngameObject::getRenderDepth() const {
+float IngameObject::getRenderDepth() const {
     return renderDepth_;
 }
